refactor :: 디렉토링 수정
