+++ application-infrastructure/src/main/kotlin/hs/kr/equus/application/global/feign/HeaderConfiguration.kt
@@ -1,4 +1,4 @@
-package hs.kr.equus.application.global.feign.config
+package hs.kr.equus.application.global.feign
 
 import feign.RequestInterceptor
 import org.springframework.context.annotation.Bean
